PYIC-8692 refactor geContraIndicatorCredentialHandler for maintainability
Efforts towards readability of the main handler, and extracting
CI types for use in other lambdas in the same stub.
